multiply very bing numbers

+++ 0x0C-more_malloc_free/holberton.h
@@ -10,7 +10,7 @@ int *array_range(int min, int max);
 void *_realloc(void *ptr, unsigned int old_size, unsigned int new_size);
 int _putchar(char c);
 int _atoi(const char *s);
-void print_int(int n);
+void print_int(unsigned long int n);
 void _puts(char *str);
 
 #endif
